+++ CHANGELOG.md
@@ -5,9 +5,15 @@
 ## [Unreleased]
 
 ### Added
-- If commit's subject line contains mentions of issues, which are configured for
+- If a commit's subject line contains mentions of issues, which are configured for
   [navigation in the project](https://www.jetbrains.com/help/idea/settings-version-control-issue-navigation.html),
   then the reference pasted into a rich text editor will include clickable links.
+- [LICENSE.txt](https://github.com/rybak/intellij-copy-commit-reference/blob/main/LICENSE.txt)
+  is now included as part of the plugin's jar file.
+
+### Fixed
+- Warning caused by incorrect usage of `com.intellij.openapi.progress.ProgressIndicator`
+  has been fixed.
 
 ## [0.3.0-alpha] - 2023-07-07