+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## [Unreleased]
 
 ### Added
-- Added "Commit Commit Reference" action to the context menu of VCS annotations.
+- Added "Copy Commit Reference" action to the context menu of VCS annotations.
 
 ## [1.0.0] - 2023-07-10